# `AveragePrecisionRecall._evaluate_image`: Don't recompute on each `min_iou`.