#### What does this implement/fix? Explain your changes.

Small refactor of `AveragePrecisionRecall._evaluate_image` to remove unnecessary recomputations.

#### Any other comments?

I was just going through the AP implementation and noticed it.

Didn't run any exhaustive benchmarks as it felt obvious. For the following snippet:

```python
from deepchecks.vision.datasets.detection.coco import load_dataset as load_coco_dataset
from deepchecks.vision.metrics_utils.detection_precision_recall import ObjectDetectionAveragePrecision
coco_test_visiondata = load_coco_dataset(train=False, object_type='VisionData', shuffle=False)
metric = ObjectDetectionAveragePrecision(return_option=None)

%%timeit
for batch in coco_test_visiondata:
    label = coco_test_visiondata.batch_to_labels(batch)
    prediction = coco_test_visiondata.infer_on_batch(batch, mock_trained_yolov5_object_detection, device)
    metric.update((prediction, label))
```

This branch:

```console
464 ms ± 26.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

main:

```console
671 ms ± 31.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```